refactor(calendar): refactoring code
