+++ packages-web/calendar/src/utils/style.ts
@@ -1,0 +1,17 @@
+export const parseStyle = (style = ""): { [key: string]: string } => {
+    try {
+        return style.split(";").reduce<{ [key: string]: string }>((styleObject, line) => {
+            const pair = line.split(":");
+            if (pair.length === 2) {
+                const name = pair[0].trim().replace(/(-.)/g, match => match[1].toUpperCase());
+                styleObject[name] = pair[1].trim();
+            }
+
+            return styleObject;
+        }, {});
+    } catch (error) {
+        console.error("Failed to parse style", style, error);
+    }
+
+    return {};
+};
